Making changes to the project for Unit test, Creating Test file, Adding mock storage. Adding necessary Packages and modifications for babelrc
