+++ src/index.js
@@ -4,18 +4,45 @@ import TaskList from './modules/Taskslist.js';
 const tasklist = new TaskList();
 
 function onPageLoad() {
-  tasklist.loadTasks();
   if (localStorage) {
-    const localStorageItem = localStorage.getItem('tasklist');
+    const localStorageItem = localStorage.getItem('tasklists');
     tasklist.readTasks(JSON.parse(localStorageItem));
+    loadTasks();
   }
+}
+
+function loadTasks() {
+  const tasks = document.querySelector('.tasks');
+  tasks.innerHTML = '';
+  tasklist.tasklist.forEach((task) => {
+    const containertask = document.createElement('div');
+    const tasklabel = document.createElement('div');
+    const taskinput = document.createElement('input');
+    const taskdescription = document.createElement('input');
+    const trashicon = document.createElement('i');
+    containertask.setAttribute('class', 'task-container');
+    tasklabel.setAttribute('class', 'currenttask');
+    taskinput.setAttribute('type', 'checkbox');
+    taskinput.setAttribute('class', 'check-box');
+    taskdescription.setAttribute('type', 'input');
+    taskdescription.setAttribute('class', 'taskfield');
+    trashicon.setAttribute('class', 'fa-solid fa-trash-can delete-icon');
+    trashicon.setAttribute('id', `removetask${task.index}`);
+    taskdescription.value = task.description;
+    taskinput.checked = task.completed;
+    tasklabel.appendChild(taskinput);
+    tasklabel.appendChild(taskdescription);
+    tasklabel.appendChild(trashicon);
+    containertask.appendChild(tasklabel);
+    tasks.appendChild(containertask);
+  });
 }
 
 document.addEventListener('DOMContentLoaded', () => {
   window.onload = onPageLoad();
 
   window.addEventListener('beforeunload', () => {
-    localStorage.setItem('tasklist', JSON.stringify(tasklist));
+    localStorage.setItem('tasklists', JSON.stringify(tasklist));
   });
 
   document.querySelector('#addtask').addEventListener('click', () => {
@@ -23,6 +50,7 @@ document.addEventListener('DOMContentLoaded', () => {
       const index = tasklist.tasklist.length;
       tasklist.addTask(document.querySelector('#taskdescription').value, false, index);
       document.querySelector('#taskdescription').value = '';
+      loadTasks();
     }
   });
 
@@ -32,6 +60,7 @@ document.addEventListener('DOMContentLoaded', () => {
         const index = tasklist.tasklist.length;
         tasklist.addTask(document.querySelector('#taskdescription').value, false, index);
         document.querySelector('#taskdescription').value = '';
+        loadTasks();
       }
     }
   });
@@ -41,6 +70,7 @@ document.addEventListener('DOMContentLoaded', () => {
     icons.forEach((button, index) => {
       if (event.target === button) {
         tasklist.removeTask(index);
+        loadTasks();
       }
     });
   });
@@ -72,5 +102,6 @@ document.addEventListener('DOMContentLoaded', () => {
         i -= 1;
       }
     }
+    loadTasks();
   });
 });
