+++ webpack.config.js
@@ -12,10 +12,8 @@ module.exports = {
     }),
   ],
   output: {
-    filename: 'main.js',
     path: path.resolve(__dirname, 'dist'),
-    clean: true,
-    publicPath: '/',
+    filename: 'main.js',
   },
   module: {
     rules: [
